Move test helper to correct file
