+++ event-store/src/test_helpers.rs
@@ -2,6 +2,9 @@ use crate::aggregator::Aggregator;
 use crate::event::Event;
 use crate::pg::PgQuery;
 use postgres::types::ToSql;
+use r2d2::Pool;
+use r2d2_postgres::{PostgresConnectionManager, TlsMode};
+use std::time::{SystemTime, UNIX_EPOCH};
 
 /// Set of all events in the domain
 #[derive(Events, Debug)]
@@ -42,4 +45,65 @@ impl Aggregator<TestEvents, String, PgQuery> for TestCounterEntity {
 
         PgQuery::new("select * from events", params)
     }
+}
+
+fn current_time_ms() -> u64 {
+    let start = SystemTime::now();
+    let since_the_epoch = start
+        .duration_since(UNIX_EPOCH)
+        .expect("Time went backwards");
+
+    let in_ms =
+        since_the_epoch.as_secs() * 1000 + since_the_epoch.subsec_nanos() as u64 / 1_000_000;
+
+    in_ms
+}
+
+/// Create a new database with a random name, returning the connection
+pub fn pg_create_random_db() -> Pool<PostgresConnectionManager> {
+    let db_id = format!("eventstorerust-test-{}", current_time_ms());
+
+    println!("Create test DB {}", db_id);
+
+    let manager =
+        PostgresConnectionManager::new("postgres://postgres@localhost:5430", TlsMode::None)
+            .unwrap();
+
+    let pool = r2d2::Pool::new(manager).unwrap();
+
+    let conn = pool.get().unwrap();
+
+    conn.batch_execute(&format!("CREATE DATABASE \"{}\"", db_id))
+        .unwrap();
+
+    let manager = PostgresConnectionManager::new(
+        format!("postgres://postgres@localhost:5430/{}", db_id),
+        TlsMode::None,
+    )
+    .unwrap();
+
+    let pool = r2d2::Pool::new(manager).unwrap();
+
+    let conn = pool.get().unwrap();
+
+    conn.batch_execute(
+        r#"
+        CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
+
+        CREATE TABLE events (
+            id uuid DEFAULT uuid_generate_v4() PRIMARY KEY,
+            data jsonb NOT NULL,
+            context jsonb DEFAULT '{}'::jsonb
+        );
+
+        CREATE TABLE aggregate_cache (
+            id VARCHAR(64) PRIMARY KEY,
+            data jsonb NOT NULL,
+            time timestamp without time zone DEFAULT now()
+        );
+    "#,
+    )
+    .unwrap();
+
+    pool
 }
