+++ event-store/tests/save_and_aggregate.rs
@@ -6,9 +6,6 @@ extern crate pretty_env_logger;
 use event_store::*;
 use futures::future;
 use futures::prelude::*;
-use r2d2::Pool;
-use r2d2_postgres::{PostgresConnectionManager, TlsMode};
-use std::time::{SystemTime, UNIX_EPOCH};
 use tokio_core::reactor::Core;
 
 #[test]
@@ -51,65 +48,4 @@ fn save_and_aggregate() {
         });
 
     core.run(run).unwrap();
-}
-
-fn current_time_ms() -> u64 {
-    let start = SystemTime::now();
-    let since_the_epoch = start
-        .duration_since(UNIX_EPOCH)
-        .expect("Time went backwards");
-
-    let in_ms =
-        since_the_epoch.as_secs() * 1000 + since_the_epoch.subsec_nanos() as u64 / 1_000_000;
-
-    in_ms
-}
-
-/// Create a new database with a random name, returning the connection
-pub fn pg_create_random_db() -> Pool<PostgresConnectionManager> {
-    let db_id = format!("eventstorerust-test-{}", current_time_ms());
-
-    println!("Create test DB {}", db_id);
-
-    let manager =
-        PostgresConnectionManager::new("postgres://postgres@localhost:5430", TlsMode::None)
-            .unwrap();
-
-    let pool = r2d2::Pool::new(manager).unwrap();
-
-    let conn = pool.get().unwrap();
-
-    conn.batch_execute(&format!("CREATE DATABASE \"{}\"", db_id))
-        .unwrap();
-
-    let manager = PostgresConnectionManager::new(
-        format!("postgres://postgres@localhost:5430/{}", db_id),
-        TlsMode::None,
-    )
-    .unwrap();
-
-    let pool = r2d2::Pool::new(manager).unwrap();
-
-    let conn = pool.get().unwrap();
-
-    conn.batch_execute(
-        r#"
-        CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
-
-        CREATE TABLE events (
-            id uuid DEFAULT uuid_generate_v4() PRIMARY KEY,
-            data jsonb NOT NULL,
-            context jsonb DEFAULT '{}'::jsonb
-        );
-
-        CREATE TABLE aggregate_cache (
-            id VARCHAR(64) PRIMARY KEY,
-            data jsonb NOT NULL,
-            time timestamp without time zone DEFAULT now()
-        );
-    "#,
-    )
-    .unwrap();
-
-    pool
 }
